buildEnv: Prefer local builds
For things like the system path, there is almost no advantage in
building on a remote machine since it's only a bunch of links that are
created.
Even worse, the remote builder will have to substitute all packages just
for symlinking them into the env. Building this on the local machine
works around this because the local machine will have to have all
packages anyway.