nixos/plausible: Fix nixpkgs-introduced ecto_sql incompatibility.
Fixes runtime error during `migrate.sh`:

    ** (UndefinedFunctionError) function ClickhouseEcto.lock_for_migrations/3 is undefined or private

The function `lock_for_migrations` indeed does not exist in the `ClickhouseEcto`
module packaged so far.

Reason:

So far we use the patch `ecto_sql-fix.patch` doing

    -      {:ecto_sql, "~> 3.0"},
    +      {:ecto_sql, "~> 3.6"},

@scvalex found that the commit that makes the `plausible/clickhouse_ecto` fork
compatible with 3.6 is _newer_ than the commit pinned in the `.lock` file,
namely  `1969f14ecef - Update for ecto 3.6.1`:

    https://github.com/plausible/clickhouse_ecto/commit/1969f14ecef

That commit introduces the function that my error shows (`def lock_for_migrations`).

This means that the version that's in the lockfile (and pulled) was
incompatible with the version the nixpkgs patch forces.

This commit fixes it by patching the `.lock` file to have the version of
`plausible/clickhouse_ecto` that makes it compatible with `ecto_sql` 3.6.