kernel: enable MOUSE_PS2_VMMOUSE
Turns VMware guest mouse support on in the kernel. This is needed for running Wayland and non-root X in a VMWare guest. In a pre-Wayland world the `xf86-input-vmmouse` userspace driver would have handled this for us. This allows the mouse to properly work in a vmware guest (for example it can now leave the vmware window).

See: https://github.com/vmware/open-vm-tools/issues/528