Change link and hash for linux64
Changed line 97 and 98 for the new download link for the 64 bit linux version of the tor broser bundle. You might want to check of I got the sha 256 hash correct, I have not done this before :). And the hash seams longer than I thought it should be. Also Long live the tor project!!