librsync_0_9: drop
This was restored in a0dcd49bbc7 ("btar: fix build by reintroducing
librsync 0.9.x for ZHF") to fix the build of a single package.  We now
use a patch to make it work with more recent versions of librsync, so
there's no reason to keep this package around any more.