kitty: 0.21.2 -> 0.23.1
https://github.com/kovidgoyal/kitty/releases/tag/v0.23.1
Add the new dependencies required for building the documentation.
Also run the postInstall hook at the end of the installPhase instead of in the middle.
The URL of the changelog has changed slightly. The old one still works because of a redirect but it's better to use the new one directly.