pympress: 1.5.1 -> 1.6.3
This also makes the application ready to work on Darwin (but still not
working until #113777 is fixed).

Version 1.5.1 did not build on Darwin because libcanberra-gtk3 is
broken, and python-vlc is required but only works on Linux.

Version 1.6.x fixed the python-vlc requirement to make it optional, so
now both libcanberra-gtk3 and python-vlc inputs can be made optional,
and enabled by default on Linux. This means that on Darwin there is no
support for media, but it is still better than not having the
application at all.

Unfortunately, until watchdog 2.x is fixed, the Darwin build will be
still broken.

-----

The setup.py file now parses the setuptools version, so it needs to be
patched to fix an error with the 'post0' suffix in the version.

Also, setuptools is needed in propagatedBuildInputs to fix the
`ModuleNotFoundError: No module named 'pkg_resources'` error.