libertinus: 6.9 -> 7.040
Besides the version bump, the Nix expression was updated
to conform to directory structure changes in the archive file,
to skip installing the documentation files,
and to use fetchurl to get the archive with the binary files
(rather than the source archive that's pulled by fetchFromGitHub).

Also, the short and long descriptions were updated
to match the current contents of the project's README.

Finally, the homepage and owner fields were changed
to reflect the current location of the project.