lib/modules: Default to emptyValue of types if no definitions
This patch makes it so that if an option has no definitions, but its
type specifies an `emptyValue`, that is used instead of throwing an error.

This means that for e.g. `attrsOf`, there's no need to specify a default
of `default = {}` anymore. Other types that have such a default are
`nullOr` (default `null`), `submodule` (default `{}`) and `listOf`
(default `[]`).