boost-build: allow supplying a boost version to build b2 for
The `useBoost` argument expects an attribute set with an src and version
attribute (so a boost derivation works) and builds b2 for the given
version of boost. This is useful when bootstrapping boost: We can
override boost-build to get a boost-build derivation matching the
current boost version we want to build (b2 is not backwards compatible
enough to build older boost versions) without the need of having
one boost-build${version} attribute for every boost version we have
which is not very useful for nixpkgs' users.