boost*: use packaged b2 instead it of building in configurePhase
If we build two things in one derivation, it becomes more complicated
and its build time is extended. Therefore we should avoid this if
possible. There's a good opportunity for this with boost: We have
boost-build packaged already. This has the additional benefit that
we can get rid of $CC_FOR_BUILD entirely in boost, meaning we don't
need to rely on (as many) hacks to make boost understand our way of
cross compiling.

Unfortunately boost-build is not backwards compatible, so we need to
build a specific boost-build for each boost derivation (the number
could probably be reduced, but I'm not interested in testing a lot
of boost builds at the moment).

Additionally we fix a few cross compilation problems:

- boost couldn't cope with different types of compilers for native
  and cross (as happens if useLLVM is true). Since we only use one
  of them per derivation, this is no longer an issue.

- boost didn't find the cross ar and ranlib for compilation (since
  it doesn't check $AR or $RANLIB). Instead it used plain ar and
  ranlib form $PATH which were the native ones before. This is now
  fixed by setting these tools explicitly in user-config.jam (and
  no longer providing the native tools).

With these changes, pkgsLLVM.boost builds.

On darwin, instead of patching the clang-darwin.jam definition, we
instead supply -rpath $out/lib via <linkflags> which causes the
correct directory to be added to the libraries' rpaths, so that
they find each other.