cabal2nix: 2.17.0 -> 2.18.0
https://github.com/NixOS/cabal2nix/blob/v2.18.0/CHANGELOG.md

2.18.0 mostly brings bug fixes, especially the one fixing the unclear
error message if a cabal file was missing (which was forgotten in the
changelog). It is not technically a non-breaking change since the
Haskell API changed.

Since there shouldn't be any breakage for users and all known reverse
dependencies of cabal2nix have been broken for a while, the benefits
of backporting outweigh the risks.