hqplayerd: install main configuration file
While packaging I had missed this in the file listing, and ended up
having to add a hack to the NixOS module to create an empty
configuration file.

Turns out that wasn't needed at all, as long as we distribute the real
configuration file with the package, which this change does.