nixos/systemd-boot: fix undefined reference to `profile`
`profile` is undefined at this point; however it is part of the `gen` tuple, so
access it through that. A `None` is converted to the string `"system"` to align
with the conventions of the rest of the script: `None` profiles refer to the
system profile, whereas anything else is some other profile by the specified
name.