haskellPackages.gtk2hs-buildtools: attempt to fix on aarch64-linux
We employ a workaround for a GHC bug [1] which has been adopted by both
Debian and Fedora for a eerily similar problem on ppc64le. Hopefully
this fixes our aarch64 issue as well (untested so far). -O0 is not
ideal, but compilation with -fllvm fails when linking (due to an
invalid relocation and passing -fPIC wasn't enough to fix it), so
we're stuck with this for now.

[1]: https://gitlab.haskell.org/ghc/ghc/-/issues/17203