nixos/luksroot: mask device UUID in boot prompts
When decrypting a partition at boot time, NixOS displays this
passphrase prompt:

    Passphrase for ${dev.device}:

...where "${dev.device}" may evaluate to, for example, "/dev/sda2",
"/dev/disk/by-partlabel/<label>", or "/dev/disk/by-partuuid/<uuid>".

The `by-partuuid` option is almost always preferrable, as it uniquely
identifies the partition that we wish to decrypt.  In which case our
boot prompt will look like this:

    Passphrase for /dev/disk/by-partuuid/f89e123d-84d3-4e62-aaa8-4fd648665b20:

There are two issues with this situation:

1. The prompt is long and unsightly
2. More importantly, it correlates a boot-time screen capture to the
   user's hard drive. That is, the UUID of a disk partition will be
   visible at boot time to bystanders who may be able to see the
   user's screen. Meanwhile, that UUID exists unencrypted on disk. An
   attacker who holds the encrypted disk as well as a screenshot of
   the boot prompt can prove that the disk belongs to the user.

The solution used here is to provide a new configuration option,
${dev.deviceDisplayName}, which will be displayed on-screen in
place of ${dev.device} during boot. (The ${dev.device} option must
still be provided as usual to identify the partition.)