haskellPackages.SDL-ttf: fix build on darwin
SDL_main.h which is included from SDL_ttf.h contains a #define which redefines
main to SDL_main. This has to do with the way SDL binaries are linked and
executed on darwin which is explained here [1]. However this is bad news for
hsc2hs intermediate C sources which in this case only to gather version
information from the headers and not to run a full SDL application.

Thus compilation would fail due to this issue which we can avoid by redefining
main to main again after including the header.

[1]: http://www.ariis.it/link/repos/sdl-bindings/hssdl-image/MACOSX