haskellPackages.ats-format: ignore custom Setup.hs, install manually
ats-format has a custom Setup.hs which does the following using cli-setup:

* Add ~/.local/share/man to $MANPATH by editing shell configurations in $HOME.

* Install a man page into ~/.local/share

* Install a shell completion by editing shell configurations in $HOME.

We had a workaround to keep this from failing the build in the nix sandbox, but
this seemingly broke on darwin. To fix this once and forall, we force the use
of only `defaultMain` in Setup.hs and install the man page manually.
Additionally we generate completions and remove unnecessary extra references
from the closure.