nixos/nullmailer: Create "failed" directory
Nullmailer expects that this directory exists (see
https://github.com/bruceg/nullmailer/blob/073f4e9c5d015ed50074afe603e7faa32e34599e/doc/nullmailer-send.8#L185).
When it doesn't and an email cannot be sent due to a permanent failure
or has been in the queue longer than queuelifetime (7 days), message
"Can't rename file: No such file or directory" starts appearing in the
log and nullmailer never sends "Could not send message" notification.
This means that the user may never learn that his email was not
delivered.