vieb: fix internal node_modules structure
Vieb's source code manually constructs the path the 7z binary using
a relative path. This breaks because vieb is running under
libexec/vieb/deps/vieb (and expects 7z to be inside
libexec/vieb/deps/vieb/node_modules), whereas the actual binary is
in libexec/vieb/node_modules). We fix this with a symlink.