custom-sysroot: replace preferLocalBuild with runCommandLocal
the core motivation was to get rid of phases.
phases itself is a whitelisting procedure when building packages.
this has one major issue: forgetting fixupPhase

there is a new approach by blacklisting every single phase, that is not needed. e.g. dontConfigure = true;

i could use the blacklisting here, but this would produce a lot of lines to prevent the usage of phases. Therefore I did some thinking and cam up with the runCommandLocal, which does nearly the same thing like preferLocalBuild` and additionally executes the few lines of code.

So overall the process does not really change, except there are fewer lines and phases is removed.