octoprint: 1.5.3 -> 1.6.1
Hardcode path to pip so octoprint will find it and recognize it cannot
write to its install dir with the /nix/store. This causes less, at least
when looking at the log output, compared to not having pip accessible at
all.

Hardcore path to ffmpeg, because it is not reasonable on a NixOS setup
to not treat this as a dependency. At the same time we hide the related
settings form, so the user cannot overwrite it.
Having the user input a sensible path would require installing ffmpeg
into the environment, not a great solution, so patching it is.