zeroad: make sure hydra builds zeroad-unwrapped
Hydra hasn't built zeroad since the 0.0.24 bump:
https://hydra.nixos.org/job/nixos/trunk-combined/nixpkgs.zeroad.x86_64-linux/all
This is due to the data package becoming larger than the output
limit. Even though the data package is already marked with
`hydraPlatforms = [];`, hydra still tried to build it, since it is a
dependency of `zeroad`. This makes it so `zeroad` isn't built by hydra
either. Only `zeroad-unwrapped` (which takes significant compilation
time) will be built by hydra.