nixos/pam: add extraConfig
Currently it is only possible to set the options in services to enable certain modules or to override the entire PAM config to load other modules that don't have a flag.

This option makes it possible to append to the autogenerated config, thus allowing to use both the flags and a custom config with custom modules at the same time.