lvm2: remove probably-unnecessary Musl patch
e606e4d6a92 ("lvm2: musl patches from alpine") added all these patches
from Alpine to fix Musl builds, but one doesn't actually seem related
to Musl.  The Alpine issue[1] that led to its introduction links to a
Gentoo issue[2] about the same thing on Glibc.

We shouldn't be applying a fix for a non-libc-specific bug only to
Musl builds, and if this was going to be an issue for us we'd expect
to have seen it on Glibc by now.  It's more likely that it's been
fixed in the meantime.

[1]: https://gitlab.alpinelinux.org/alpine/aports/-/issues/3107
[2]: https://bugs.gentoo.org/335492