conda: invoke bash login shell to source /etc/profile
The `profile` parameter in `buildFHSUserEnv` is used to create an /etc/profile file
which should be sourced by bash. However, the default runScript is a non-login bash
which does not source /etc/profile. Therefore the conda package breaks, as a script
which is required for `conda activate`, is not sourced. Also the installationPath
parameter is ignored as it is setup in the /etc/profile file.