haskell.compiler.ghcHEAD: increase Hydra timeout because Darwin builds were timing out
Building `haskell.compiler.ghcHEAD` on Hydra on Darwin seemed to cause
frequent timeouts.

Here's an example of a timeout: https://hydra.nixos.org/build/149762652

This commit raises the timeout time on Hydra for ghcHEAD.  The other
GHC derivations have this set, so it makes sense to set this here
as well.