i2p: get rid of duplicate Java Service Wrapper
The definition of Java Service Wrapper in i2p is identical to the java-service-wrapper package except that the latter also executes the test suite, which requires some extra dependencies.
The java-service-wrapper package was actually added in https://github.com/NixOS/nixpkgs/pull/67824 to move the definition from the i2p package into a separate package. While it was added as a separate package, it was never removed from i2p. This commit fixes that.