candle: write user settings to ~/.config/candle, not ${candle}/bin
Candle attempts to write user-specific settings into the application
directory (`${candle}/bin`), which is obviously a non-starter for Nix.
Worse, the inability to write settings causes a bad initial UI state,
where the path-visualization area is completely black (because all the
customizable colors, including the background, are incorrectly set to
black). The colors must be readjusted on every execution of the program.

This patch sends the settings to `~/.config/candle/settings.ini`
instead, and thereby solves the all-black display issue. (In the
unlikely case that `$HOME` is not set, `HOME=/tmp` is used as a
fallback.)